Refactor getHorarioDisponible y getListaEventoByDiaAndEmpresaId
