+++ src/main/kotlin/com/estonianport/agendaza/service/EventoService.kt
@@ -7,8 +7,13 @@ import com.estonianport.agendaza.model.Empresa
 import com.estonianport.agendaza.model.Evento
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.data.repository.CrudRepository
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Service
+import org.springframework.util.CollectionUtils
 import java.time.LocalDateTime
+import java.util.stream.Collectors
+import java.util.stream.IntStream
 
 @Service
 class EventoService : GenericServiceImpl<Evento, Long>() {
@@ -37,7 +42,84 @@ class EventoService : GenericServiceImpl<Evento, Long>() {
         return empresa.listaEvento.any{ it.codigo == codigo}
     }
 
-    fun findAllByStartdBetweenAndEmpresa(inicio: LocalDateTime, fin: LocalDateTime, empresa: Empresa): List<Evento> {
+    fun findAllByInicioBetweenAndListaEmpresa(empresa: Empresa, desde: LocalDateTime, hasta: LocalDateTime): List<Evento> {
+
+        // Crea la hora inicio y la hora final de un dia para buscar todos los eventos en X dia
+        val inicio: LocalDateTime = LocalDateTime.of(desde.year, desde.month, desde.dayOfMonth, 0,0 )
+        val fin: LocalDateTime = LocalDateTime.of(desde.year, desde.month, desde.dayOfMonth, 23,59 )
+
         return eventoDao.findAllByInicioBetweenAndListaEmpresa(inicio, fin, empresa)
+    }
+
+    fun getHorarioDisponible(listaEvento: List<Evento>, desde : LocalDateTime, hasta : LocalDateTime) : Boolean{
+        
+        // En caso de no existir ningun evento para esa fecha devolver disponible
+        if (!listaEvento.isEmpty()) {
+
+            // lista de todas las lista de rangos horarios
+            val listaDeRangos: MutableList<List<Int>> = ArrayList()
+
+            // Variable usada para obtener la hora final del evento
+            var horaFinal = ""
+
+            // Obtiene el rango horario de los eventos agendados
+            for (evento in listaEvento) {
+                if (evento.inicio.plusDays(1).getDayOfMonth() == evento.fin.getDayOfMonth()) {
+                    horaFinal = suma24Horas(evento.fin)
+                } else {
+                    horaFinal = evento.fin.toLocalTime().toString()
+                }
+                listaDeRangos.add(getRango(evento.inicio.toLocalTime().toString(), horaFinal))
+            }
+
+            // Obtiene el rango horario del nuevo evento a agendar
+            if (desde.dayOfMonth != hasta.dayOfMonth) {
+                horaFinal = suma24Horas(hasta)
+            } else {
+                horaFinal = hasta.toLocalTime().toString()
+            }
+            val rangoEventoNuevo = getRangoConMargen(desde.toLocalTime().toString(), horaFinal)
+
+            // Si intercepta algun rango de hora
+            for (rangos in listaDeRangos) {
+                if (CollectionUtils.containsAny(rangos, rangoEventoNuevo)){
+                    return false
+                }
+            }
+        }
+        // Si no intercepta a ninguna da el disponible
+        return true
+    }
+    
+    private fun getRango(inicio: String, fin: String): List<Int> {
+        val horaInicioSplit = inicio.split(":")
+        val horaFinSplit = fin.split(":")
+
+        val horaInicio = (horaInicioSplit[0] + horaInicioSplit[1]).toInt()
+        val horaFin = (horaFinSplit[0] + horaFinSplit[1]).toInt()
+
+        return IntStream.range(horaInicio, horaFin).boxed().collect(Collectors.toList())
+    }
+
+    private fun getRangoConMargen(inicio: String, fin: String): List<Int> {
+        val horaInicioSplit = inicio.split(":")
+
+        val horaFinSplit = fin.split(":")
+
+        var horaInicio = (horaInicioSplit[0] + horaInicioSplit[1]).toInt()
+        var horaFin = (horaFinSplit[0] + horaFinSplit[1]).toInt()
+
+        // Le agrega una hora antes y una hora despues para tener margen
+        horaInicio -= 100
+        horaFin += 100
+        return IntStream.range(horaInicio, horaFin).boxed().collect(Collectors.toList())
+    }
+
+    private fun suma24Horas(fechaFin: LocalDateTime): String {
+        var horaFin: String = fechaFin.toLocalTime().hour.toString()
+        val minutosFin: String = fechaFin.toLocalTime().minute.toString()
+        val finHoraEventos = horaFin.toInt() + 24
+        horaFin = Integer.toString(finHoraEventos)
+        return "$horaFin:$minutosFin"
     }
 }
